% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module-Dev.R
\name{ChainLadder}
\alias{ChainLadder}
\alias{BornhuetterFerguson}
\alias{BF}
\alias{CapeCod_ELR}
\alias{CapeCod}
\title{Methods to estimate ultimates for a triangle}
\usage{
ChainLadder(x, link_ratio = NULL, tail_factor = 1)

BornhuetterFerguson(x, expected_loss, link_ratio = NULL, tail_factor = 1)

BF(x, expected_loss, link_ratio = NULL, tail_factor = 1)

CapeCod_ELR(x, premium, link_ratio = NULL, tail_factor = 1)

CapeCod(x, premium, link_ratio = NULL, tail_factor = 1)
}
\arguments{
\item{x}{A Triangle object}

\item{link_ratio}{A numeric vector of link ratios}

\item{tail_factor}{A numeric number for tail factor}

\item{expected_loss}{A numeric vector of expected losses, used for B-F method}

\item{premium}{A numeric vector of premiums, used for Cape-Cod method}
}
\description{
Methods to estimate ultimates for a triangle
}
\examples{
loss <- Triangle(TRI$Incurred)
ata <- tri_LinkRatioAverage(loss, "simple")
premium <- seq(10000, 28000, by = 2000)

# Chain Ladder
ChainLadder(loss, link_ratio = ata)

# B-F
ELoss_BF <- c(5000, 6000, 7000, 8000, 9000, 8000, 8800, 9600, 10400, 11200)
BornhuetterFerguson(loss, ELoss_BF, link_ratio = ata)

# Cape-Cod
CapeCod(loss, premium, ata)
CapeCod_ELR(loss, premium, ata) ## CapeCod ELR that is used in `CapeCod` function call
}
