% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exhibit-ggplot-helper.R
\name{color_contrast}
\alias{color_contrast}
\title{calculate contrast color for better visualization}
\usage{
color_contrast(color)
}
\description{
calculate contrast color for better visualization
}
\examples{
cors <- cor(mtcars[, 1:4])

# Melt matrix
df <- data.frame(
  col = colnames(cors)[as.vector(col(cors))],
  row = rownames(cors)[as.vector(row(cors))],
  value = as.vector(cors)
)

# Basic plot
p <- ggplot(df, aes(row, col, fill = value)) +
  geom_raster() +
  geom_text(aes(label = round(value, 2), color = after_scale(color_contrast(fill))))
p + scale_fill_viridis_c(direction =  1)
}
