% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exhibit-ggplot-helper.R
\name{color_palette}
\alias{color_palette}
\title{A selection of color palettes}
\usage{
color_palette(name = c("economist", "gc"), idx = NULL)
}
\description{
A selection of color palettes
}
\examples{
color_palette() |> plot()
color_palette("economist", c(1, 3, 5, 6)) |> plot()
}
