% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-Data-wrangle.R
\name{data.trable}
\alias{data.trable}
\title{Similar to \code{tibble::trible} to create \code{data.table} using row-by-row input}
\usage{
data.trable(...)
}
\arguments{
\item{...}{Arguments specifying the structure of a \code{data.table}.
Variable names should be formulas, and may only appear before the data.}
}
\description{
Similar to \code{tibble::trible} to create \code{data.table} using row-by-row input
}
\examples{
data.trable(
~colA, ~colB,
"a",   1,
"b",   2,
"c",   3
)
}
