% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module-DataTest.R
\name{DataTest}
\alias{DataTest}
\title{A container for all data tests}
\description{
A container for all data tests

A container for all data tests
}
\examples{
tester <- DataTest$new()
DT <- data.table::as.data.table(mtcars)
tester$test_that("cyl is below 20", {DT[cyl > 20]})
tester$test_that("cyl cannot be 6", {DT[cyl == 6]}, "cyl2")
## or wrap them all into a function
test_f <- function() {
  tester <- DataTest$new()
  tester$test_that("cyl is below 20", {DT[cyl > 20]})
  tester$test_that("cyl cannot be 6", {DT[cyl == 6]}, "cyl2")
}
test_f()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataTest-new}{\code{DataTest$new()}}
\item \href{#method-DataTest-test_that}{\code{DataTest$test_that()}}
\item \href{#method-DataTest-summary}{\code{DataTest$summary()}}
\item \href{#method-DataTest-write_to_excel}{\code{DataTest$write_to_excel()}}
\item \href{#method-DataTest-print}{\code{DataTest$print()}}
\item \href{#method-DataTest-keep_cols}{\code{DataTest$keep_cols()}}
\item \href{#method-DataTest-clone}{\code{DataTest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTest-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataTest-new}{}}}
\subsection{Method \code{new()}}{
Create a new DataTest object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTest$new(cols_keep = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols_keep}}{A character vector for columns to keep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTest-test_that"></a>}}
\if{latex}{\out{\hypertarget{method-DataTest-test_that}{}}}
\subsection{Method \code{test_that()}}{
Create a new test
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTest$test_that(
  desc,
  code,
  test_name = NULL,
  cols_forward = NULL,
  cols_keep = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{desc}}{Human-readble description for the test}

\item{\code{code}}{Code for testing, can be embraced with curly bracket (i.e. {...})}

\item{\code{test_name}}{Concise test id, usually for outputing as Excel sheetname.  If left blank, an id like "test-001" will be created automatically}

\item{\code{cols_forward}}{Columns to bring forward in the result dataset}

\item{\code{cols_keep}}{Columns in output dataset (for invalid records).  This will add to the \code{cols_keep} input during object initialization}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTest-summary"></a>}}
\if{latex}{\out{\hypertarget{method-DataTest-summary}{}}}
\subsection{Method \code{summary()}}{
Summarise the results for all tests
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTest$summary(invalid_only = TRUE, sorted = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{invalid_only}}{Logical, only show invalid tests}

\item{\code{sorted}}{Logical, whether to reorder the tests by \code{test_name}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTest-write_to_excel"></a>}}
\if{latex}{\out{\hypertarget{method-DataTest-write_to_excel}{}}}
\subsection{Method \code{write_to_excel()}}{
Output test summary and invalid datasets to Excel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTest$write_to_excel(file, headerStyle = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{File path for excel output}

\item{\code{headerStyle}}{Not implemented yet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTest-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataTest-print}{}}}
\subsection{Method \code{print()}}{
For S3 print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTest-keep_cols"></a>}}
\if{latex}{\out{\hypertarget{method-DataTest-keep_cols}{}}}
\subsection{Method \code{keep_cols()}}{
Overwrite keep_cols on the flight
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTest$keep_cols(cols_keep = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols_keep}}{A character vector for columns to keep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataTest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
