% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataType-extension.R
\name{dots_list}
\alias{dots_list}
\title{Work on \code{...}}
\usage{
dots_list(...)
}
\description{
Work on \code{...}
}
\examples{
dots_list(a = 1, b = head(mtcars))
dots_list(list(a = 1, b = head(mtcars))) # when the 1st and only component is a list
add_spec <- function(x, ...) c(x, dots_list(...))
add_spec(list(a = 1), b = 2, c = 3)
}
