% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-Data-wrangle.R
\name{dt_patch}
\alias{dt_patch}
\title{Patch new data y to old dataset x}
\usage{
dt_patch(x, y, by, vars = NULL, ties = c("y", "x"))
}
\arguments{
\item{x}{Data to be patched}

\item{y}{Patching dataset}

\item{by}{A character vector specifying common join ids}

\item{vars}{A character vector specifying common fields to be patched}

\item{ties}{Specify what to do when both \code{x} and \code{y} has valid values for \code{vars}}
}
\description{
Essentially a safer way of doing left_join
}
\examples{
dt1 <- data.table::data.table(ID1 = c(1, 1, 2, 2, 3), ID2 = c("A", "B", "A", "C", "B"), v1 = 1:5, v2 = 5:1)
dt2 <- data.table::data.table(ID1 = c(1, 2, 3), ID2 = "A", v1 = 11L, v2 = 22L)
dt_patch(dt1, dt2, by = c("ID1", "ID2"), vars = c("v1"))
dt_patch(dt1, dt2, by = c("ID1", "ID2"), vars = c("v1", "v2"))
}
