% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-Data-wrangle.R
\name{dt_unite}
\alias{dt_unite}
\title{By-ref equivalent of \code{tidyr::unite}}
\usage{
dt_unite(DT, cols = NULL, into, sep = "_", remove = TRUE, na.rm = TRUE)
}
\arguments{
\item{DT}{A \code{data.table} or \code{data.frame}.}

\item{cols}{A char vector containing the colnames to be united.}

\item{into}{The name of the new column, as a string.}

\item{sep}{Separator to use between values.}

\item{remove}{If \code{TRUE}, remove input columns from output data frame.}
}
\description{
By-ref equivalent of \code{tidyr::unite}
}
\examples{
df <- as.data.table(expand.grid(x = c("a", NA), y = c("b", NA)))
dt_unite(copy(df), c("x", "y"), into = "new")[]
dt_unite(copy(df), c("x", "y"), into = "new", sep = "...", remove = FALSE)[]
dt_unite(copy(df), into = "new")[]
dt_unite(copy(df), into = "new", na.rm = TRUE)[]
}
