% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module-ExposureRating.R
\name{ExposureRating}
\alias{ExposureRating}
\alias{fMean.ExposureRating}
\alias{build_CDF}
\title{Create an ExposureRating module}
\usage{
ExposureRating(RiskProfile, ExpectLossRatio, RiskCurve, SubjectPremium = NULL)

## Loss in Layer
fMean(ExposureRating)

build_CDF(x, q = NULL)
}
\description{
Create an ExposureRating module
}
\examples{
x <- data.table(Lbound = c(0, 100, 200), Ubound = c(100, 200, Inf), N = 10, TSI = c(50, 150, 250) * 10, Premium = c(50, 150, 250) * 10 * .001)
RP <- RiskProfile(x)
PropExpoRating <- ExposureRating(RP, ExpectLossRatio = .45, SubjectPremium = 9, RiskCurve(b = 0.0218339035107131, g = 1096.98894971097))

## Mean Loss
## ... is passed to XL function
fMean(PropExpoRating, Limit = c(50, 200), Deductible = c(50, 100))
fMean(PropExpoRating, c("50 xs 50", "200 xs 100"))

## CDF
build_CDF(PropExpoRating, mseq(50, 250, n = 200))
}
