% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataType-extension.R
\name{field_map}
\alias{field_map}
\title{Work on fields of a \code{data.frame}}
\usage{
field_map(x, ...)
}
\description{
Work on fields of a \code{data.frame}
}
\examples{
DT <- data.frame(A = 1:3, B = LETTERS[5:7])
data.table::copy(DT) |> field_map(Num = "A", Char = "B")
data.table::copy(DT) |> field_map(c(Num = "A", Char = "B")) |> field_check("Char") |> field_fill(c("X", "Y"))
}
