% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Generic.R
\name{fPerc}
\alias{fPerc}
\alias{fPerc.default}
\title{Generic function for percentile (i.e. inverse of Value at Risk)}
\usage{
fPerc(x, ...)

\method{fPerc}{default}(x, q)
}
\description{
Generic function for percentile (i.e. inverse of Value at Risk)
}
