% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-FileSystem.R
\name{fs_common_dir}
\alias{fs_common_dir}
\title{Create common directory structure for analysis}
\usage{
fs_common_dir(path = NULL, flavor = c("op", "gc"), create_project = TRUE)
}
\description{
Create common directory structure for analysis
}
