% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Generic.R
\name{fTVaR}
\alias{fTVaR}
\alias{fTVaR.default}
\title{Generic function for Tail Value at Risk (a.k.a Conditional Tail Expectation, or CTE)}
\usage{
fTVaR(x, ...)

\method{fTVaR}{default}(x, prob)
}
\description{
Generic function for Tail Value at Risk (a.k.a Conditional Tail Expectation, or CTE)
}
