% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Generic.R
\name{fVaR}
\alias{fVaR}
\alias{fVaR.default}
\title{Generic function for Value at Risk (i.e. quantile)}
\usage{
fVaR(x, ...)

\method{fVaR}{default}(x, prob)
}
\description{
Generic function for Value at Risk (i.e. quantile)
}
