% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exhibit-ggplot-helper.R
\name{gg_save}
\alias{gg_save}
\title{output a ggplot chart}
\usage{
gg_save(plt, file, width = 9, height = 6, scale = 1)
}
\arguments{
\item{plt}{Plot object}

\item{file}{File name (with path)}

\item{width, height}{Width and height in inches}

\item{scale}{Scale factor for \code{width} and \code{height}.  If \code{scale} is 2, and \code{width} is 10, then scaled width will be 20}
}
\description{
output a ggplot chart
}
