% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-Data-wrangle.R
\name{makenm_numbered}
\alias{makenm_numbered}
\alias{makenm_unique}
\alias{makenm_clean}
\title{Make names from a character vector}
\usage{
makenm_numbered(x, prefix = "V", blank_only = FALSE)

makenm_unique(x, prefix = "V")

makenm_clean(
  x,
  prefix = "V",
  sep = "_",
  style = c("asis", "lower", "title", "upper"),
  allow_na = FALSE,
  allow_dup = FALSE
)
}
\arguments{
\item{prefix}{Prefix of numbered names, e.g. "v" as a prefix for "v1".}

\item{blank_only}{Only change names for blank items, i.e. keep original non-NA names.}

\item{style}{Case style of names, will change the case of prefix if applicable.}

\item{allow_na}{If FALSE, will replace \code{NA}s to v-prefix names, e.g. "v1".}

\item{allow_dup}{If FALSE, duplicated occurrence will be labeled. e.g. "Var", "Var2", "Var3".  Note that it is case-sensitive so "var" and "Var" are viewed as different names.}
}
\description{
Make names from a character vector

\code{makenm_clean} is inspired by \code{janitor::make_clean_names} and \code{snakecase::to_any_case}
}
\examples{
x <- c(NA, "Gross Premium ", "Loss --- Incurred", "Premium/Limit", "# of Risks", "Adj. Rate (\%)", NA, "GrossPremium")
makenm_numbered(x)
makenm_numbered(x, prefix = "...")
makenm_numbered(x, blank_only = TRUE)
makenm_unique(c("A", "B", "A", NA))
makenm_clean(x)
makenm_clean(x, style = "title")
makenm_clean(x, sep = "")
makenm_clean(x, style = "asis", sep = "", allow_na = TRUE, allow_dup = TRUE)
## A more common use of makenm functions are for reading data into tibbles using packages like readr or readxl
list(LETTERS[1:5], 1:5) |> tibble::as_tibble(.name_repair = makenm_numbered)
}
