% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-MetaRisk.R
\name{mr_read_OccurrenceTable}
\alias{mr_read_OccurrenceTable}
\title{Read MetaRisk By-Occurrence detailed data}
\usage{
mr_read_OccurrenceTable(file, ...)
}
\description{
Read MetaRisk By-Occurrence detailed data
}
