% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-string.R
\name{num}
\alias{num}
\alias{num_format}
\title{Format numbers}
\usage{
num(
  x,
  scale = c("auto", "C", "K", "M", "B", "P", "pp"),
  digits = 2,
  na.label = "N/A",
  inf.label = "Inf",
  ...
)

num_format(scale = c("auto", "C", "K", "M", "B", "P", "pp"), digits = 0, ...)
}
\description{
Format numbers
}
\examples{
x <- c(1.23e12, 3.45e7, -6.78e3, Inf, NA)
num(x)
num(x, "M", digits = 1)
num(x, "B", inf.label = "Unlimited", na.label = "")
}
