% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataType-OccurrenceTable.R
\name{OccurrenceTable}
\alias{OccurrenceTable}
\title{Occurrence Table}
\usage{
OccurrenceTable(
  x,
  Realization = "Realization",
  OccurrenceId = "OccurrenceId",
  EventId = "EventId"
)
}
\description{
Occurrence Table
}
\examples{
set.seed(123)
DT <- data.table::CJ(Year = 1:100000, OccId = 1:2, EventId = NA)[, `:=`(x = runif(200000), y = runif(200000))]
YOT <- OccurrenceTable(DT, "Year", "OccId", "EventId")
fMean(YOT)
fMean(YOT, "OEP")
fSD(YOT)
fPerc(YOT, c(1.000291, 1.550170, 1.859187))
fVaR(YOT, c(.5, .9, .99), "AEP")
fVaR(YOT, c(.5, .9, .99), "OEP")
fTVaR(YOT, c(.5, .9, .99))
fMax(YOT)
}
