% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-string.R
\name{parse_numeric}
\alias{parse_numeric}
\title{Parse a character vector to numeric}
\usage{
parse_numeric(x, na = c("", "NA", "N/A"))
}
\description{
Parse a character vector to numeric
}
\examples{
parse_numeric(c("$123,456.789", "N/A", "3.1415926525K", "99.95\%", "Inf", "inf"))
}
