% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-ContractTerm.R
\name{parse_XL}
\alias{parse_XL}
\alias{parse_reinstatement}
\title{Helper function to create a XL contract structure}
\usage{
parse_XL(
  x = NULL,
  Limit = Inf,
  Deductible = 0,
  AAL = Inf,
  AAD = 0,
  Reinstatement = NA_character_,
  Premium = 0
)

parse_reinstatement(reinstat)
}
\arguments{
\item{reinstat}{a character atomic variable, e.g. "2@100\\%"}
}
\description{
Helper function to create a XL contract structure

Parse reinstatement terms into data.frame
}
\examples{
## parse various XL terms
parse_XL(c("100M xs 200M", "200M xs 300M"))
parse_XL(c("100M xs 200M xs 50M", "200M xs 300M"))
parse_XL(Limit = 100, Deductible = 200, Reinstatement = "2@100\%")
parse_XL(Limit = 100, Deductible = c(100, 200))
## parse various reinstatement terms
parse_reinstatement("1@120\%, 2@100\%")
parse_reinstatement("1@120\%;2@100\%")
parse_reinstatement(c(1.2, 1))
parse_reinstatement(0) ## note this mean 1@0\%, rather than Nil reinstatement
parse_reinstatement("2@120\%")
parse_reinstatement("2@free")
parse_reinstatement("uf")
parse_reinstatement("nil")
parse_reinstatement("0@100\%")
}
