% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantre-package.R
\docType{package}
\name{quantre-package}
\alias{quantre}
\alias{quantre-package}
\title{quantre: An R toolkit for Quantitative Analysis in Reinsurance}
\description{
`quantre` is an R toolkit designed to accelerate modeling, pricing and risk-management process in reinsurance transactions
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/addnox/quantre}
  \item Report bugs at \url{https://github.com/addnox/quantre/issues}
}

}
\author{
\strong{Maintainer}: Charles Pan \email{addnox@gmail.com}

}
\keyword{internal}
