% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-Excel-read.R
\name{readxl}
\alias{readxl}
\alias{readxl_raw}
\alias{readxl_named_range}
\title{DT version of read_excel}
\usage{
readxl(
  wb,
  ws = NULL,
  range = NULL,
  rows = NULL,
  cols = NULL,
  col_names = TRUE,
  col_types = NULL,
  ...
)

readxl_raw(
  wb,
  ws = NULL,
  range = NULL,
  rows = NULL,
  cols = NULL,
  col_names = FALSE,
  col_types = "text",
  alpha_type_cols = FALSE,
  ...
)

readxl_named_range(wb, ws = NULL)
}
\description{
DT version of read_excel

Read entire Excel sheet from cell \code{A1}

Read named ranges in workbook
}
