% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-Excel-read.R
\name{readxl_dfc}
\alias{readxl_dfc}
\title{Read ranges and cbind the results}
\usage{
readxl_dfc(
  wb,
  ws = NULL,
  range,
  col_names = TRUE,
  clean_names = TRUE,
  sep = NULL
)
}
\description{
Read ranges and cbind the results
}
