% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-Excel-read.R
\name{readxl_vector}
\alias{readxl_vector}
\title{Read one column or row into vector}
\usage{
readxl_vector(wb, ws = NULL, range, data_type = c("text", "date", "numeric"))
}
\description{
Read one column or row into vector
}
