% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module-Pricer.R
\name{ReportOutput}
\alias{ReportOutput}
\title{Create a ReportOutput object}
\usage{
ReportOutput(x, ...)
}
\description{
Create a ReportOutput object
}
\examples{
set.seed(123)
DT <- data.table::data.table(Year = 1:10, x = runif(10, max = 2), y = runif(10, max = 10))
ReportOutput(DT, Realization = "Year", Loss = "y") ## note that x is deleted, and y is kept as Loss

}
