% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataType-RiskProfile.R
\name{RiskProfile}
\alias{RiskProfile}
\alias{fScale.RiskProfile}
\title{Create a RiskProfile object}
\usage{
RiskProfile(x, ...)

## Scale a Risk Profile
fScale(RiskProfile, k)
}
\description{
Create a RiskProfile object
}
\examples{
x <- data.frame(N = c(100, 50, 10), SumInsured = c(100 * 10, 50 * 20, 10 * 30), Prem = c(10, 10, 10))
rp <- RiskProfile(x, TSI = "SumInsured", Premium = "Prem")
rp

## Scale a Risk Profile
fScale(rp, 3)
fScale(rp, c(3, 2, 0))
}
