% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-string.R
\name{stri_amap}
\alias{stri_amap}
\title{Approximate string mapping}
\usage{
stri_amap(x, y)
}
\description{
Approximate string mapping
}
\examples{
x1 <- c("水稻种植保险\\\\nRice Insurance", "水地玉米种植保险\\\\nIrrigated Land Corn Insurance", "旱地玉米种植保险\\\\nDryland Corn Insurance", "水地小麦种植保险\\\\nIrrigated Land Wheat Insurance ", "旱地小麦种植保险\\\\nDryland Wheat Insurance", "水地马铃薯保险\\\\nIrrigated Potato Insurance ", "旱地马铃薯保险\\\\nDryland Potato Insurance", "油菜种植保险\\\\nRape Insurance")
x2 <- c("Irrigated Land Corn Insurance (水地玉米种植保险)", "Rice Insurance (水稻种植保险)", "Irrigated Land Wheat Insurance（水地小麦种植保险）", "Rape Insurance （油菜种植保险)", "Dryland Potato Insurance (旱地马铃薯保险)", "Dryland Wheat Insurance（旱地小麦种植保险）", "Irrigated Potato Insurance (水地马铃薯保险)", "Dryland Corn Insurance (旱地玉米种植保险)")
data.frame(Orig = x1, Mapped = stri_amap(x1, x2))
}
