% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-string.R
\name{stri_cj}
\alias{stri_cj}
\title{Cross-join character vectors and paste}
\usage{
stri_cj(..., sep = "_")
}
\arguments{
\item{...}{One or more character vectors, as in \code{paste} function}

\item{sep}{A character string to separate the terms}
}
\description{
Cross-join character vectors and paste
}
\examples{
stri_cj(c("SI", "Prem"), c("Gross", "SP", "Net"), c("Actual", "AsIf"), sep = "..")
}
