% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-string.R
\name{stri_date_parse}
\alias{stri_date_parse}
\title{Parse a character vector to Date}
\usage{
stri_date_parse(x)
}
\description{
Parse a character vector to Date
}
\examples{
x <- c("2021-01-03", "12-31-2019", "Jan 2 2022", "2023/1/1")
stri_date_parse(x)
}
