% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-string.R
\name{stri_full_to_half}
\alias{stri_full_to_half}
\title{Convert full-width characters into half-width}
\usage{
stri_full_to_half(x)
}
\description{
Convert full-width characters into half-width
}
\examples{
x <- c("(Irrigated Land Corn InsuranceABC,.?!) （水地玉米种植保险ＡＢＣ，。？！）", "Rice Insurance (水稻种植保险)")
stri_full_to_half(x)
}
