% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-string.R
\name{stri_numeric_parse}
\alias{stri_numeric_parse}
\title{Parse a character vector to numeric}
\usage{
stri_numeric_parse(x, na = c("", "NA", "N/A"))
}
\description{
Parse a character vector to numeric
}
\examples{
stri_numeric_parse(c("$123,456.789", "N/A", "3.1415926525K", "99.95\%", "Inf", "inf"))
}
