% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-string.R
\name{stri_trunc}
\alias{stri_trunc}
\alias{stri_squish}
\title{Truncate a string to maximum width}
\usage{
stri_trunc(x, width, side = c("right", "left", "center"), ellipsis = "...")

stri_squish(x)
}
\arguments{
\item{x}{Input vector. Either a character vector, or something coercible to one.}

\item{width}{Maximum width of string.}

\item{side, ellipsis}{Location and content of ellipsis that indicates content has been removed.}
}
\description{
Code borrowed and modified from \code{stringr} package
}
\examples{
x <- "This string is moderately long"
rbind(
  stri_trunc(x, 20, "right"),
  stri_trunc(x, 20, "left"),
  stri_trunc(x, 20, "center")
)
}
