% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-string.R
\name{stri_wrap0}
\alias{stri_wrap0}
\title{Wrap words into nicely formatted paragraphs}
\usage{
stri_wrap0(string, width = 80, indent = 0, exdent = 0, whitespace_only = TRUE)
}
\description{
Wrap words into nicely formatted paragraphs
}
