% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataType-OccurrenceTable.R
\name{sum_across_var}
\alias{sum_across_var}
\title{Sum across multiple columns}
\usage{
sum_across_var(x, var = -(1:3))
}
\arguments{
\item{x}{An instance of \code{OccurrenceTable} object}

\item{var}{Either a numeric vector of positions, or a character vector of names}
}
\description{
Sum across multiple columns
}
