% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-tidy_table.R
\name{tidy_header}
\alias{tidy_header}
\title{Escape version for tidy_header}
\usage{
tidy_header(
  DT,
  long_cols = NULL,
  wide_cols = NULL,
  wide_names = NULL,
  wide_fill = 1:(nrow(DT) - 1),
  wide_split = NULL,
  cols_delete = NULL,
  cols_keepnames = NULL,
  clean_long_names = TRUE,
  verbose = FALSE
)
}
\description{
Escape version for tidy_header
}
