% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-tidy_helper.R
\name{tidy_hsplit}
\alias{tidy_hsplit}
\alias{tidy_vsplit}
\title{Split data into list of \code{data.table}s horizontally or vertically}
\usage{
tidy_hsplit(x, by = NULL, na.rm = FALSE, keep.by = TRUE, factor.by = FALSE)

tidy_vsplit(x, by = NULL, na.rm = TRUE, keep.by = TRUE, factor.by = FALSE)
}
\arguments{
\item{x}{A data.frame}

\item{by}{Numeric vector indicating position}

\item{na.rm}{When TRUE, those rows with all \code{cols} being \code{NA} will be deleted before splitting}
}
\description{
Split data into list of \code{data.table}s horizontally or vertically
}
\examples{
DT1 <- data.table(ID1 = c(NA, "A", "A", NA, NA, "B"), ID2 = c(rep("1", 3), NA, "2", "2"), X1 = NA, X2 = c(NA, 1:2, NA, NA, 3), X3 = c(rep(NA, 5), .1))
tidy_hsplit(DT1)
tidy_hsplit(DT1, by = 1:2)
tidy_vsplit(data.table::transpose(DT1))

DT2 <- data.table(ID1 = c(1, NA, NA, 2, NA), ID2 = c("A", NA, "B", "C", NA), X1 = 1:5, X2 = LETTERS[1:5])
tidy_hsplit(DT2, by = 1:2)

}
