% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-tidy_helper.R
\name{tidy_trim}
\alias{tidy_trim}
\title{Trim all-blank row and cols of a data.table}
\usage{
tidy_trim(DT)
}
\arguments{
\item{DT}{A \code{data.table}}
}
\description{
Trim all-blank row and cols of a data.table
}
\examples{
DT <- data.table(A = c(NA, 1, 2, NA, NA, 3), B = NA, C = c(NA, "A", NA, NA, NA, "B"), D = NA)
tidy_trim(DT)
}
