% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-tidy_helper.R
\name{tidy_where}
\alias{tidy_where}
\alias{tidy_which}
\title{Return row index (or logical vector) that match the pre-defined patterns}
\usage{
tidy_where(DT, patterns, exclusions = NULL)

tidy_which(DT, patterns, exclusions = NULL, as.DT = TRUE)
}
\arguments{
\item{patterns}{A vector means "AND" relations. See examples.}

\item{exclusions}{Patterns to be excluded (i.e. set to FALSE in results).}
}
\description{
Return row index (or logical vector) that match the pre-defined patterns
}
\examples{
x <- data.table(
  c("ABC P&C Limited", "Layer", "Limit", NA, 10, 10, 10, "Total Limit"),
  c(NA, NA, "Excess", NA, 10, 20, 30, "")
)
tidy_where(x, "limit")
tidy_where(x, c("limit|excess"), exclusions = "ABC")
tidy_which(x, c("limit|excess"), exclusions = "ABC")
}
