% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exhibit-ggplot-helper.R
\name{transform_dual_axis}
\alias{transform_dual_axis}
\title{Helper function for dual-y-axis plot}
\usage{
transform_dual_axis(y_left, y_right, force_zero = FALSE)
}
\arguments{
\item{y_left, y_right}{Numeric vector.
If with two elements, linear transformation will be used
If longer than two, min/max (or zero/max, if \code{force_zero} is \code{TRUE}) are used for linear transformation}
}
\description{
Helper function for dual-y-axis plot
}
\examples{
a <- data.table(Year = 2021:2023, Value = c(10, 12, 9), Growth = c(-.1, .2, -.25))
trans1 <- transform_dual_axis(c(10, 20), c(0, .2))
trans2 <- transform_dual_axis(a$Value, a$Growth)
trans3 <- transform_dual_axis(a$Value, a$Growth, force_zero = TRUE)
ggplot(a, aes(Year, Value)) +
geom_col() +
geom_line(aes(y = trans2$to_left(Growth))) +
scale_y_continuous(sec.axis = sec_axis(transform = trans2$transform))

}
