% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataType-Triangle.R
\name{tri_block}
\alias{tri_block}
\title{Split triangle into pairs of development years for further average link ratio calculation}
\usage{
tri_block(x, window = Inf, exclude = 0)
}
\description{
Split triangle into pairs of development years for further average link ratio calculation
}
