% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataType-Triangle.R
\name{Triangle}
\alias{Triangle}
\alias{tri_LatestDiagonal}
\alias{tri_FullTriangle}
\alias{tri_LinkRatio}
\alias{tri_LinkRatioSummary}
\alias{tri_LinkRatioAverage}
\alias{tri_ReportedPercent}
\alias{tri_CDF}
\title{\code{Triangle}: Create a triangle object}
\usage{
Triangle(x, from = 12, by = 12, latest = from)

tri_LatestDiagonal(x)

tri_FullTriangle(x, report_pct, ultimate_loss = NULL)

tri_LinkRatio(x)

tri_LinkRatioSummary(x)

tri_LinkRatioAverage(
  x,
  type = c("weighted", "simple", "geometric"),
  window = Inf,
  exclude = 0
)

tri_ReportedPercent(link_ratio, tail_factor = 1)

tri_CDF(link_ratio, tail_factor = 1)
}
\description{
\code{Triangle}: Create a triangle object

\code{tri_LatestDiagonal}: Get latest diagonal from a Triangle

\code{tri_FullTriangle}: Develope and fill bottom-right part of the triangle

\code{tri_LinkRatio}: Get age-to-age link ratio for each origin row

\code{tri_LinkRatioSummary}: Quick summary of commonly used link ratio averages

\code{tri_LinkRatioAverage}: Calculate average link ratio factors for development

\code{tri_ReportedPercent}: Convert age-to-age link ratios to reported percents

\code{tri_CDF}: Convert age-to-age link ratios to cumulative development factors
}
\examples{
loss <- Triangle(TRI$Incurred)

## two triangles can do arithmetic
loss / loss
loss * 2

# get latest diagonal
tri_LatestDiagonal(loss)

# Take a look at different averages for link ratio
tri_LinkRatioSummary(loss)

# Use selected link ratios to develope a triangle
ata <- tri_LinkRatioAverage(loss, "weighted", 5) ## weighted avg. for latest 5 years
reportPct <- tri_ReportedPercent(ata, tail_factor = 1.02)
tri_FullTriangle(loss, reportPct)
}
