% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-string.R
\name{vdetect}
\alias{vdetect}
\alias{vpick}
\title{Search for pattern}
\usage{
vdetect(
  x,
  patterns = NULL,
  exclusions = NULL,
  offset = 0L,
  negate = FALSE,
  ignore_case = TRUE,
  na.rm = TRUE
)

vpick(
  x,
  patterns = NULL,
  exclusions = NULL,
  y = x,
  offset = 0L,
  fill_direction = NULL,
  negate = FALSE,
  ignore_case = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{x}{A character vector.}

\item{patterns}{A vector means "AND" relations. See examples.}

\item{exclusions}{Patterns to be excluded (i.e. set to FALSE in results).}

\item{offset}{Offset the results (see sample).}

\item{negate}{single logical value; whether a no-match to \code{patterns} is rather of interest}

\item{na.rm}{Set \code{NA}s in original data as FALSE.}

\item{ignore.case}{Decide whether the \code{patterns} are case-insensitive.}
}
\description{
Search for pattern
}
\examples{
x <- c(NA, NA, "layer", "no", 1, 2, 3, "layer info:")
x[vdetect(x, c("layer|no"))] # OR relation by regular expression
x[vdetect(x, c("layer|no"), na.rm = FALSE)]
x[vdetect(x, c("layer", "info"))] # AND relation for vector
x[vdetect(x, c("layer|no"), exclusions = "info")]
x[vdetect(x, c("layer$|no"))]
x[vdetect(x, c("layer$|no"), negate = TRUE)]
## offset
y <- c(NA, NA, "2020", "(Jan-1)", "Company A", NA, NA, "2018/19", "0701", "Comp A")
y[vdetect(y, "20\\\\d{2}")] ## easier to get year info
y[vdetect(y, "20\\\\d{2}", offset = 1)] ## use offset for inconsistent month info
y[vdetect(y, "^comp", offset = -1)] ## negative offset means moving upwards
}
