% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-string.R
\name{vlookup}
\alias{vlookup}
\title{Excel \code{vlookup} equivalent}
\usage{
vlookup(x, key, value = key, ignore_case = TRUE, default = NULL, ...)
}
\description{
Excel \code{vlookup} equivalent
}
\examples{
(x <- c("A", "B", "C", "CBA", "abc"))
key <- c("a", "b", "BA")
value <- c("Alice", "Bob", "Apple")
vlookup(x, key, value)
vlookup(x, key, value, ignore_case = FALSE)
vlookup(x, key, value, default = "Nobody")
vlookup(x, key, value, default = x)
}
