% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-string.R
\name{vnafill}
\alias{vnafill}
\title{A more general \code{nafill} that can work with non-numeric data type}
\usage{
vnafill(
  x,
  direction = c("constant", "down", "up", "forwards", "backwards"),
  fill = NA,
  breaks = NULL
)
}
\arguments{
\item{direction}{Direction in which to fill missing values.
Currently either "down" (the default), "up", "downup" (i.e. first down and then up) or "updown" (first up and then down).
"forwards" is an alias of "down", as "backwards" is of "up".}
}
\description{
A more general \code{nafill} that can work with non-numeric data type
}
\examples{
x <- c("A", NA, "B", NA, NA, NA, "C", "D", NA)
vnafill(x, "down")
vnafill(x, "up")
vnafill(x, fill = "missing")
vnafill(x, "down", breaks = c(F, F, F, F, T, F, F, F, F))
}
