% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-Excel-read.R
\name{xlrange}
\alias{xlrange}
\title{Parse Excel range information}
\usage{
xlrange(range = NULL, rows = NULL, cols = NULL, ws = NULL)
}
\description{
Parse Excel range information
}
\examples{
xlrange("A1")
xlrange("A1:B3", ws = "Data")
xlrange(rows = 2:10)
xlrange(rows = c(3, NA)) # NA means to page end
xlrange(cols = "A:C")
xlrange(rows = 2:5, cols = c(9, NA))
}
