% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataType-OccurrenceTable.R
\name{OccurrenceTable}
\alias{OccurrenceTable}
\title{Occurrence Table}
\usage{
OccurrenceTable(
  x,
  Realization = "Realization",
  OccurrenceId = "OccurrenceId",
  EventId = "EventId",
  Value = NULL
)
}
\description{
Occurrence Table
}
\examples{
set.seed(123)
DT <- data.table::CJ(Year = 1:100000, OccId = 1:2, EventId = NA)[, `:=`(x = runif(200000), y = runif(200000))]
OccTbl <- OccurrenceTable(DT, "Year", "OccId", "EventId")
OEPTbl <- EP(OccTbl, "OEP")
fMean(OccTbl)
fMean(OccTbl, "OEP")
fMean(OEPTbl)
fSD(OccTbl)
fPerc(OccTbl, c(1.000291, 1.550170, 1.859187))
fVaR(OccTbl, c(.5, .9, .99), "AEP")
fVaR(OccTbl, c(.5, .9, .99), "OEP")
fTVaR(OccTbl, c(.5, .9, .99))
fMax(OccTbl)
}
